Add assignment files
Added all assignment files. conversation_prompt.py is a repurposed version of
the file by the same name from the original assignment with slight alteration
to the handle_conversation_data function and a new function, search_document,
for searching a dictionary of links related to different topics. The links
are contained in a json document, documentation_links.json.